backend/requirements.txt
flask
torch
transformers
janus
whisper
jinja2
pillow